invest contracts refactor
