+++ src/invest/invest-deployed-contracts/invest-deployed-contracts.model.ts
