+++ src/invest/invest-staking-steps/invest-staking-steps.tsx
@@ -24,7 +24,7 @@ import {
 import * as deployModel from '~/automations/automation-deploy-contract/automation-deploy-contract.model'
 import * as automationUpdateModel from '~/automations/automation-update/automation-update.model'
 import * as model from '~/invest/invest-detail/invest-detail.model'
-import * as stakingAutomatesModel from '~/staking/staking-automates/staking-automates.model'
+import * as stakingAutomatesModel from '~/invest/invest-deployed-contracts/invest-deployed-contracts.model'
 import * as stakingAdaptersModel from '~/staking/staking-adapters/staking-adapters.model'
 import * as lpTokensModel from '~/lp-tokens/lp-tokens.model'
 import * as styles from './invest-staking-steps.css'
