+++ src/invest/invest-unstaking-steps/invest-unstaking-steps.tsx
@@ -12,7 +12,7 @@ import { InvestSell } from '~/invest/invest-sell'
 import { InvestUnstakingStepsUnstake } from './invest-unstaking-steps-unstake'
 import { InvestUnstakingStepsSuccess } from './invest-unstaking-steps-success'
 import { useQueryParams } from '~/common/hooks'
-import * as stakingAutomatesModel from '~/staking/staking-automates/staking-automates.model'
+import * as stakingAutomatesModel from '~/invest/invest-deployed-contracts/invest-deployed-contracts.model'
 import * as stakingAdaptersModel from '~/staking/staking-adapters/staking-adapters.model'
 import * as model from '~/invest/invest-detail/invest-detail.model'
 import * as lpTokensModel from '~/lp-tokens/lp-tokens.model'
