+++ src/lp-tokens/lp-tokens.tsx
@@ -22,7 +22,7 @@ import { walletNetworkModel } from '~/wallets/wallet-networks'
 import { networksConfig } from '~/networks-config'
 import { LPTokensSell } from './lp-tokens-sell'
 import { LPTokensBuySellDialog } from './common/lp-tokens-buy-sell-dialog'
-import * as stakingAutomatesModel from '~/staking/staking-automates/staking-automates.model'
+import * as stakingAutomatesModel from '~/invest/invest-deployed-contracts/invest-deployed-contracts.model'
 import * as stakingModel from '~/staking/staking-adapters/staking-adapters.model'
 import { settingsWalletModel } from '~/settings/settings-wallets'
 import { LPContracts } from './common/lp-tokens.types'
