+++ src/staking/staking-adapters/staking-adapters.tsx
@@ -22,7 +22,7 @@ import { toastsService } from '~/toasts'
 import { settingsWalletModel } from '~/settings/settings-wallets'
 import { BlockchainEnum } from '~/api/_generated-types'
 import { GovernanceStake, GovernanceUnstake } from '~/common/load-adapter'
-import * as stakingAutomatesModel from '~/staking/staking-automates/staking-automates.model'
+import * as stakingAutomatesModel from '~/invest/invest-deployed-contracts/invest-deployed-contracts.model'
 import * as model from './staking-adapters.model'
 import * as styles from './staking-adapters.css'
 import { LPTokensBuySellDialog } from '~/lp-tokens/common/lp-tokens-buy-sell-dialog'
