--- src/staking/staking-automates/index.ts
@@ -1,1 +1,0 @@
-export * from './staking-automates'
